+++ msgctl.go
@@ -6,7 +6,7 @@ import (
 
 // Msgctl calls the msgctl() syscall.
 // FIXME: we are not passing the buf argument, see msgctl(2).
-func Msgctl(qid uint64, cmd int) error {
+func Msgctl(qid uint, cmd int) error {
 	var buf uintptr = 0
 	_, _, err := syscall.Syscall(syscall.SYS_MSGCTL, uintptr(qid), buf, 0)
 	if err != 0 {
